==> football_data_analysis/1.get_tables_groupstage.ipynb <==
import pandas as pd
import pickle
from string import ascii_uppercase as alphabet

all_tables = pd.read_html('https://en.wikipedia.org/wiki/2022_FIFA_World_Cup')

print(all_tables[11])

print(all_tables[18])

print(all_tables[25])

for i in range(11,18):
  print(all_tables[i])
  print("\n")

for i in range(19,31):
  print(all_tables[i])
  print("\n")

for i in range(32,45):
  print(all_tables[i])
  print(i)
  print("\n")

for i in range(39,46):
  print(all_tables[i])
  print(i)
  print("\n")
